+++ R/PriceIndices.R
@@ -415,6 +415,8 @@
 #'
 #' \code{\link{compare_indices_list}}
 #'
+#' \code{\link{compare_indices_jk}}
+#'
 #' \code{\link{compare_distances}}
 #'
 #' \code{\link{compare_to_target}}